+++ src/enclave/rpcendpoint.h
@@ -2,10 +2,10 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "framedtlsendpoint.h"
 #include "httpendpoint.h"
 #include "node/rpc/jsonrpc.h"
 #include "rpcmap.h"
-#include "tlsframedendpoint.h"
 
 namespace enclave
 {
