+++ src/enclave/rpcsessions.h
@@ -5,6 +5,7 @@
 #include "ds/logger.h"
 #include "ds/serialized.h"
 #include "enclavetypes.h"
+#include "framedtlsendpoint.h"
 #include "rpcclient.h"
 #include "rpcendpoint.h"
 #include "rpchandler.h"
@@ -12,7 +13,6 @@
 #include "tls/client.h"
 #include "tls/context.h"
 #include "tls/server.h"
-#include "tlsframedendpoint.h"
 
 #include <limits>
 #include <unordered_map>
